Modularized code structure.
