+++ .gitignore
@@ -20,3 +20,4 @@ Cargo.lock
 .DS_Store
 
 .idea
+.vscode
